requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
firebase-admin==6.4.0
pydantic[email]==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1
PyJWT==2.8.0 